Effective Version 1.0 of TB BIG update Context.xml is now controlling the databae configuration. Moved to Tomcat 6 for this to happen. Tomcat 5.5 on OSX has classpath issues that prevented the loading of the driver. Changed some config files to reflect new TC6 layout.
Tags are now build with StringBuilder. This makes the operation much faster and less intensive, memory and CPU wise.

Output rendering is pushed further to the custom tags (as it should be), away from the process beans.

Have a sembalence of input sanitisation. Still needs furthur work but it is rudamentaly there.

More documentation included.

Removed testbeans until proper integration in worked out (next on list)

Coupled (needs to be seperated?) in with the sanitisation is the dynamic selection builder.