Effective 0.99 of Web
Most of the structure and layout are ready.
Needs more front end design (mostly in the CSS files)

Deleted some folders under images. (Error on images/profileicons, it remains for now to try and clear the error). *Added back in after cleaning up with CL fix of checkout, delete and readd*

** Overwrote my backup to this in not the upgrade I want it to be. Still many thngs are working**