Version 1.5 Really solid release.
Updates to the stylesheet.

Archive is full functional. (Added page)
Tag now work as expected. Most uses(all?) are linked in. All tags are returned or list of articles posted under the tag.

Outbound filters are in place. Need to work out how to put into the header (does not parse JTL yet -> imported page.) and custom tags (not response model).

Important notices functionality is now in place. It displays a date based message as a notice above the articles on the front page only (could be changed if JSTL worked in imported pages.